@@ -12,37 +12,36 @@
     "prestart": "yarn",
     "start": "mastarm build --serve example.js:assets/index.js",
     "start-webpack": "webpack --watch",
-    "test": "mastarm lint",
+    "test": "mastarm lint lib/*",
     "semantic-release": "semantic-release pre && npm publish && semantic-release post"
   },
   "author": "Conveyal",
   "license": "MIT",
   "dependencies": {
-    "color": "^0.11.4",
-    "debug": "^2.3.3",
+    "@conveyal/lonlat": "^1.3.0",
+    "d3-color": "^1.0.3",
+    "debug": "^2.6.4",
     "jsolines": "^0.2.3",
-    "lodash.fill": "^3.4.0",
-    "lodash.slice": "^4.2.0",
-    "lonlng": "^0.2.0",
-    "web-worker-promise-interface": "^0.2.1"
+    "lodash": "^4.17.4",
+    "web-worker-promise-interface": "^0.3.0"
   },
   "devDependencies": {
-    "babel-loader": "^6.2.10",
-    "babel-polyfill": "^6.22.0",
-    "concat-stream": "^1.5.2",
+    "babel-loader": "^6.4.1",
+    "babel-polyfill": "^6.23.0",
+    "concat-stream": "^1.6.0",
     "isomorphic-fetch": "^2.2.1",
+    "leaflet": "^1.0.2",
     "leaflet-transitivelayer": "^0.2.0",
     "lodash.reverse": "^4.0.1",
-    "mapbox.js": "^3.0.1",
-    "mastarm": "^3.0.0",
-    "normalize.css": "^5.0.0",
-    "react": "^15.4.1",
-    "react-dom": "^15.4.1",
+    "mastarm": "^3.8.0",
+    "normalize.css": "^6.0.0",
+    "react": "^15.5.4",
+    "react-dom": "^15.5.4",
     "semantic-release": "^6.3.2",
     "tape": "^4.6.3",
     "transitive-js": "^0.9.2",
-    "webpack": "^2.2.1",
-    "webworkify-webpack": "^2.0.1"
+    "webpack": "^2.4.1",
+    "webworkify-webpack": "^2.0.5"
   },
   "standard": {
     "parser": "babel-eslint"
