@@ -23,7 +23,13 @@ const MAX_TRIP_LENGTH = 120 // minutes
  * It's also better than having duplicate code to avoid computing a surface sometimes. We could add a
  * switch to this function to select what is generated but that may even be more complexity than is needed.
  */
-export default function getSurfaceAndSpectrogramData ({origin, query, stopTreeCache, grid, which}) {
+export default function getSurfaceAndSpectrogramData ({
+  grid,
+  origin,
+  query,
+  stopTreeCache,
+  which
+}) {
   debug('generating surface and spectrogram data')
   const surface = new Uint8Array(query.width * query.height)
   const waitTimes = new Uint8Array(query.width * query.height)
@@ -42,7 +48,7 @@ export default function getSurfaceAndSpectrogramData ({origin, query, stopTreeCa
     query,
     stopTreeCache,
     origin,
-    callback ({
+    next ({
       travelTimesForDest,
       walkTimesForDest,
       inVehicleTravelTimesForDest,
