@@ -1,15 +1,12 @@
 /** Create transitive data for journeys from a particular origin to a particular destination */
 
-import Color from 'color'
-import dbg from 'debug'
-import fill from 'lodash.fill'
-import slice from 'lodash.slice'
-import ll from 'lonlng'
+import lonlat from '@conveyal/lonlat'
+import {hsl} from 'd3-color'
+import fill from 'lodash/fill'
+import slice from 'lodash/slice'
 
 import {pixelToLat, pixelToLon} from './mercator'
 import {getNonTransitTime, ITERATION_WIDTH} from './origin'
-
-const debug = dbg('browsochrones:get-transitive-data')
 
 // maximum path segments before we conclude that either R5 or Browsochrones is off its rocker and bail
 // (we do this so that we don't generate and attempt to render crazy paths; it implies a bug but it's better to log
@@ -19,16 +16,33 @@ const MAX_PATH_SEGMENTS = 8
 // arbitrary value that determines how aggressively to cluster results (higher is more aggressive, and a high number is something like 1e-3)
 const MAXIMUM_DISSIMILARITY = 5e-6
 
-export default function getTransitiveData ({ origin, query, stopTreeCache, network, from, to }) {
+export default function getTransitiveData ({
+  from,
+  log,
+  network,
+  origin,
+  query,
+  stopTreeCache,
+  to
+}) {
   // create the journeys
-  let output = {
+  const output = {
     places: [],
     journeys: []
   }
 
   // from can be left null and it will be inferred from the destination
-  const fromCoord = from != null ? ll(from) : {}
-  const fromName = from != null ? from.name : null
+  let fromCoord = {}
+  let fromName = null
+  if (from != null) {
+    const {x, y, ...rest} = from
+    try {
+      fromCoord = lonlat(rest)
+    } catch (e) {
+      log(e.message)
+    }
+    fromName = from.name
+  }
 
   output.places.push({
     place_id: 'from',
@@ -39,8 +53,13 @@ export default function getTransitiveData ({ origin, query, stopTreeCache, netwo
 
   // to cannot be undefined
   // Omit X and Y so as not to confuse lonlng
-  const { x, y, ...rest } = to // eslint-disable-line
-  const toCoord = ll(rest)
+  const { x, y, ...rest } = to
+  let toCoord = {}
+  try {
+    toCoord = lonlat(rest)
+  } catch (e) {
+    log(e.message)
+  }
 
   output.places.push({
     place_id: 'to',
@@ -51,12 +70,12 @@ export default function getTransitiveData ({ origin, query, stopTreeCache, netwo
 
   // find relevant paths at each minute
   // this array stores the destination stop and the path index in that stop
-  let { paths } = getPaths({origin, query, stopTreeCache, to})
+  let { paths } = getPaths({log, origin, query, stopTreeCache, to})
 
   paths = paths.filter(p => p !== undefined)
 
   paths.sort((p1, p2) => {
-    let v0 = p1[0] - p2[0] // sort first by stop ID
+    const v0 = p1[0] - p2[0] // sort first by stop ID
     if (v0 !== 0) return v0
     else return p1[1] - p2[1] // then by path index
   })
@@ -74,19 +93,21 @@ export default function getTransitiveData ({ origin, query, stopTreeCache, netwo
   // sort by frequency
   paths.sort((p1, p2) => p2.count - p1.count)
 
-  debug(`${paths.length} unique paths`)
+  log(`${paths.length} unique paths`)
 
   if (paths.length > 5) {
-    debug(`eliminated ${paths.length - 5} paths with frequency less than or equal to ${paths[5].count} / ${npaths} `)
+    log(`eliminated ${paths.length - 5} paths with frequency less than or equal to ${paths[5].count} / ${npaths} `)
   }
 
   paths = paths.slice(0, 5).map(p => p.path)
 
   // uniquify the paths
   // always return first path, and don't return paths the same as the previous
-  paths = paths.filter((p, idx, arr) => idx === 0 || p[0] !== arr[idx - 1][0] || p[1] !== arr[idx - 1][1]).map(p => getPath({pathDescriptor: p, origin}))
+  paths = paths.filter((p, idx, arr) =>
+    idx === 0 || p[0] !== arr[idx - 1][0] || p[1] !== arr[idx - 1][1]).map(p =>
+      getPath({log, pathDescriptor: p, origin}))
 
-  paths = filterPaths(paths, network)
+  paths = filterPaths({log, network, paths})
 
   for (let pidx = 0; pidx < paths.length; pidx++) {
     const journey = {
@@ -101,7 +122,7 @@ export default function getTransitiveData ({ origin, query, stopTreeCache, netwo
     if (path === null) continue
 
     if (path.length > 20) {
-      debug('excessive path length, more than 20 segments')
+      log('excessive path length, more than 20 segments')
       continue
     }
 
@@ -120,7 +141,7 @@ export default function getTransitiveData ({ origin, query, stopTreeCache, netwo
       })
     } else {
       // add the first walk segment
-      let originStop = path[0][0]
+      const originStop = path[0][0]
       journey.segments.push({
         type: 'WALK',
         from: {
@@ -136,19 +157,19 @@ export default function getTransitiveData ({ origin, query, stopTreeCache, netwo
       let previousStop = -1
 
       for (let segmentIdx = 0; segmentIdx < path.length; segmentIdx++) {
-        let [boardStop, pattern, alightStop] = path[segmentIdx]
+        const [boardStop, pattern, alightStop] = path[segmentIdx]
 
         // figure out from and to stop indices
-        let fromStopId = boardStop + ''
-        let toStopId = alightStop + ''
+        const fromStopId = boardStop + ''
+        const toStopId = alightStop + ''
 
         let index = 0
-        let patternData = network.patterns[pattern + '']
+        const patternData = network.patterns[pattern + '']
 
         while (patternData.stops[index].stop_id !== fromStopId) index++
-        let fromStopIdx = index
+        const fromStopIdx = index
         while (patternData.stops[index].stop_id !== toStopId) index++
-        let toStopIdx = index
+        const toStopIdx = index
 
         if (previousStop > 0 && previousStop !== boardStop) {
           // there is an on-street transfer
@@ -196,7 +217,7 @@ export default function getTransitiveData ({ origin, query, stopTreeCache, netwo
   // target is output so as not to modify network object
   Object.assign(output, network)
 
-  bundle(output)
+  bundle({data: output, log})
 
   return output
 }
@@ -253,7 +274,11 @@ export function getPaths ({origin, to, stopTreeCache, query}) {
   [from stop ID, pattern, to stop ID],
   ...
 ] */
-export function getPath ({pathDescriptor, origin}) {
+export function getPath ({
+  log,
+  origin,
+  pathDescriptor
+}) {
   const [ stop, pathIdx ] = pathDescriptor
   if (stop === -1) return []
   // offset to the first path, skipping times and nPaths
@@ -270,7 +295,7 @@ export function getPath ({pathDescriptor, origin}) {
 
   if (nSegments > MAX_PATH_SEGMENTS) {
     // bail if we have a ridiculous path (clearly a data/code error), and complain loudly
-    debug(`Too many path segments (${nSegments} > ${MAX_PATH_SEGMENTS}) in path ${pathIdx} to stop ${stop} from origin ${origin.x}, ${origin.y}, returning no path. This implies a bug, please raise cain about it`)
+    log(`Too many path segments (${nSegments} > ${MAX_PATH_SEGMENTS}) in path ${pathIdx} to stop ${stop} from origin ${origin.x}, ${origin.y}, returning no path. This implies a bug, please raise cain about it`)
     return null
   }
 
@@ -286,12 +311,16 @@ export function getPath ({pathDescriptor, origin}) {
 }
 
 /** Perform heuristic filtering on paths. How does it work? Magic. */
-export function filterPaths (paths, network) {
-  let inPathCount = paths.length
+export function filterPaths ({
+  log,
+  network,
+  paths
+}) {
+  const inPathCount = paths.length
 
   // function to convert pattern ID to route ID
-  let r = (pattern) => {
-    let pat = network.patterns.find(p => p.pattern_id === String(pattern))
+  const r = (pattern) => {
+    const pat = network.patterns.find(p => p.pattern_id === String(pattern))
     return network.routes.findIndex(r => r.route_id === pat.route_id)
   }
 
@@ -301,8 +330,8 @@ export function filterPaths (paths, network) {
     else if (p1.length > p2.length) return 1
     else {
       for (let i = 0; i < p1.length; i++) {
-        let r1 = r(p1[i][1])
-        let r2 = r(p2[i][1])
+        const r1 = r(p1[i][1])
+        const r2 = r(p2[i][1])
         if (r1 < r2) return -1
         if (r1 > r2) return 1
       }
@@ -314,7 +343,7 @@ export function filterPaths (paths, network) {
 
   paths = paths.filter((p, i, a) => {
     if (i === 0) return true
-    let prev = a[i - 1]
+    const prev = a[i - 1]
     if (p.length !== prev.length) return true
 
     for (let s = 0; s < p.length; s++) {
@@ -324,7 +353,7 @@ export function filterPaths (paths, network) {
     return false
   })
 
-  let pathCountAfterMultiTransfer = paths.length
+  const pathCountAfterMultiTransfer = paths.length
 
   // eliminate longer paths if there is a shorter path that is a subset (eliminate the short access/egress/transfer leg problem)
   paths = paths.filter((path, i, a) => {
@@ -342,7 +371,7 @@ export function filterPaths (paths, network) {
       }
 
       if (otherPathIsSubset) {
-        debug(`path ${otherPath.map(seg => r(seg[1]))} is subset of path ${routes}`)
+        log(`path ${otherPath.map(seg => r(seg[1]))} is subset of path ${routes}`)
         return false
       }
     }
@@ -350,7 +379,7 @@ export function filterPaths (paths, network) {
     return true
   })
 
-  debug(`filtering reduced ${inPathCount} paths to ${pathCountAfterMultiTransfer} after multiple-stop elimination, to ${paths.length} after stemming`)
+  log(`filtering reduced ${inPathCount} paths to ${pathCountAfterMultiTransfer} after multiple-stop elimination, to ${paths.length} after stemming`)
 
   return paths
 }
@@ -361,7 +390,7 @@ export function filterPaths (paths, network) {
  * It might also make sense to use a metric based on speed so that a very slow bus isn't bundled with a fast train,
  * but we don't currently do this.
  */
-export function bundle (data) {
+export function bundle ({data, log}) {
   // perform hierarchical clustering on journeys
   // see e.g. James et al., _An Introduction to Statistical Learning, with Applications in R_. New York: Springer, 2013, pg. 395.
   // convert to arrays
@@ -387,14 +416,12 @@ export function bundle (data) {
       }
     }
 
-    debug(`dissimilarity ${minDis}`)
+    log(`dissimilarity ${minDis}`)
+    if (minDis > MAXIMUM_DISSIMILARITY) break
 
-    // should we make a cluster?
-    if (minDis <= MAXIMUM_DISSIMILARITY) {
-      // cluster the least dissimilar clusters
-      clusters[minI] = clusters[minI].concat(clusters[minJ])
-      clusters.splice(minJ, 1) // remove clusters[j]
-    }
+    // cluster the least dissimilar clusters
+    clusters[minI] = clusters[minI].concat(clusters[minJ])
+    clusters.splice(minJ, 1) // remove clusters[j]
   }
 
   // merge journeys together
@@ -430,17 +457,18 @@ export function bundle (data) {
   // data.journeys = clusters.map(c => c[0])
 
   // color routes according to cluster
-  let col = Color('#b00')
-  clusters.forEach(c => {
-    col.rotate(67)
-    c.forEach(j => {
-      j.segments.forEach(s => {
-        s.color = col.hexString()
+  const color = hsl('#b00')
+  clusters.forEach((c) => {
+    color.h += 67
+    const rgb = color + ''
+    c.forEach((j) => {
+      j.segments.forEach((s) => {
+        s.color = rgb
       })
     })
   })
 
-  debug(`replaced ${inputSize} journeys with ${clusters.length} clusters`)
+  log(`replaced ${inputSize} journeys with ${clusters.length} clusters`)
 }
 
 /**
