@@ -1,4 +1,5 @@
 import assert from 'assert'
+import dbg from 'debug'
 import WebWorkerPromiseInterface from 'web-worker-promise-interface'
 
 import {create as createGridFunc} from './grid'
@@ -6,6 +7,8 @@ import {latToPixel, lonToPixel} from './mercator'
 
 const imageHeight = 256
 const imageWidth = 256
+
+const debug = dbg('browsochrones')
 
 export default class Browsochrones extends WebWorkerPromiseInterface {
   originLoaded = false
@@ -28,14 +31,21 @@ export default class Browsochrones extends WebWorkerPromiseInterface {
     if (transitiveNetwork) this.setTransitiveNetwork(transitiveNetwork)
   }
 
+  log (...args) {
+    debug(...args)
+  }
+
   /**
    * transfer a grid (as a raw arraybuffer) into this instance's worker. Note that this will neuter the grid, make a copy
    * before calling this function if you expect to use it elsewhere
    */
-  putGrid (key, grid) {
+  putGrid ({id, grid}) {
+    assert(typeof id === 'string', 'Id must be a string')
+    assert(grid instanceof ArrayBuffer, 'Grid must be an ArrayBuffer')
+
     return this.work({
       command: 'putGrid',
-      message: { key, grid }
+      message: {id, grid}
     })
   }
 
@@ -50,6 +60,7 @@ export default class Browsochrones extends WebWorkerPromiseInterface {
   }
 
   drawTile = (canvas, {x, y}, zoom) => {
+    assert(this.query, 'A query is required to draw a tile')
     const scaleFactor = Math.pow(2, zoom - this.query.zoom)
 
     // find top-left coords at zoom 10
@@ -75,21 +86,21 @@ export default class Browsochrones extends WebWorkerPromiseInterface {
     })
   }
 
-  async getAccessibilityForGrid (grid, cutoff = 60) {
-    assert(grid, 'A grid is required to get accessibility.')
+  async getAccessibilityForGrid ({gridId, cutoff = 60}) {
+    assert(typeof gridId === 'string', 'A grid is required to get accessibility.')
     assert(this.isLoaded(), 'Accessibility cannot be computed before generating a surface.')
 
     return this.work({
       command: 'accessibilityForGrid',
       message: {
-        grid, cutoff
+        gridId, cutoff
       }
     })
   }
 
-  async generateDestinationData ({ from, to }) {
-    assert(from, '`from` is required to generate destination data.')
-    assert(to, '`to` is required to generate destination data.')
+  async generateDestinationData ({from, to}) {
+    assert(from && from.x && from.y, '`from` point is required to generate destination data.')
+    assert(to && to.x && to.y, '`to` point is required to generate destination data.')
     assert(this.isLoaded(), 'Transitive data cannot be generated if Browsochrones is not fully loaded.')
 
     return this.work({
@@ -119,14 +130,14 @@ export default class Browsochrones extends WebWorkerPromiseInterface {
     })
   }
 
-  async generateSurface (grid, which = 'MEDIAN') {
-    assert(grid, 'A grid is required to generate a surface.')
+  async generateSurface ({gridId, which = 'MEDIAN'}) {
+    assert(gridId, 'A gridId is required to generate a surface.')
     assert(this.isLoaded(), 'Surface cannot be generated if Browsochrones is not fully loaded.')
 
     return this.work({
       command: 'generateSurface',
       message: {
-        grid,
+        gridId,
         which
       }
     }).then((message) => {
@@ -158,9 +169,10 @@ export default class Browsochrones extends WebWorkerPromiseInterface {
    * @param {Number} currentZoom
    * @returns {Object} point
    */
-  pixelToOriginPoint ({x, y}, currentZoom) {
+  pixelToOriginPoint ({pixel, currentZoom}) {
     assert(this.query, 'Cannot convert point without query.')
 
+    let {x, y} = pixel
     const {north, west, zoom} = this.query
     const scale = Math.pow(2, zoom - currentZoom)
 
@@ -186,7 +198,10 @@ export default class Browsochrones extends WebWorkerPromiseInterface {
    * @param {Number} point.x
    * @param {Number} point.y
    */
-  setOrigin (arrayBuffer, point) {
+  setOrigin ({arrayBuffer, point}) {
+    assert(arrayBuffer instanceof ArrayBuffer, 'Origin data must be an ArrayBuffer')
+    assert(!isNaN(point.x) && !isNaN(point.y), 'Point must contain an x and y value')
+
     return this.work({
       command: 'setOrigin',
       message: {
@@ -255,7 +270,7 @@ export default class Browsochrones extends WebWorkerPromiseInterface {
   }
 
   /** Get a GeoJSON isochrone with the given cutoff (in minutes) */
-  getIsochrone (cutoff = 60, interpolation = true) {
+  getIsochrone ({cutoff = 60, interpolation = true}) {
     return this.work({
       command: 'getIsochrone',
       message: {
