@@ -3,25 +3,21 @@
  * colorScheme is a function that takes a number of minutes and returns an rgba array of the color for that pixel
  */
 
-import Color from 'color'
-import slice from 'lodash.slice'
+import {hsl, rgb} from 'd3-color'
+import slice from 'lodash/slice'
 
 const colorTable = new Uint8Array(255 * 4)
 
 for (let i = 0; i < 254; i++) {
   const val = Math.min(i, 120)
+  const hue = 45 * (val / 15 | 0) // 15 minutes of travel time represents 45 degrees on the color wheel
+  const saturation = (val % 15 + 5) / 20 // saturation represents gradation within a 15-minute period
+  const luminosity = (val % 15 + 5) / 20
+  const color = rgb(hsl(hue, saturation, luminosity))
 
-  // 15 minutes of travel time represents 45 degrees on the color wheel
-  const color = Color().hsl(0, 1, 70)
-  color.rotate(45 * (val / 15 | 0))
-
-  // saturation represents gradation within a 15-minute period
-  color.saturation((val % 15 + 5) / 20 * 100)
-  color.lightness(100 - (val % 15 + 5) / 20 * 100)
-
-  colorTable[val * 4] = color.red()
-  colorTable[val * 4 + 1] = color.green()
-  colorTable[val * 4 + 2] = color.blue()
+  colorTable[val * 4] = color.r
+  colorTable[val * 4 + 1] = color.g
+  colorTable[val * 4 + 2] = color.b
   colorTable[val * 4 + 3] = 100
 }
 
