@@ -88,8 +88,8 @@ export function getNonTransitTime (origin, {x, y}) {
 
     // we can possibly walk to this pixel
     // the index of the pixel, plus one because the radius is recorded at the start of the array
-    let timeSecs = origin.data[rely * (origin.radius * 2 + 1) + relx + 1]
-    let timeMins = timeSecs / 60 | 0
+    const timeSecs = origin.data[rely * (origin.radius * 2 + 1) + relx + 1]
+    const timeMins = timeSecs / 60 | 0
     if (timeSecs !== -1 && timeMins < 255) return timeMins
   }
 
