@@ -15,7 +15,12 @@
 
 import {getNonTransitTime, ITERATION_WIDTH} from './origin'
 
-export default function propagate ({ query, stopTreeCache, origin, callback }) {
+export default function propagate ({
+  next,
+  origin,
+  query,
+  stopTreeCache
+}) {
   const {nMinutes} = origin
   const travelTimesForDest = new Uint8Array(nMinutes) // the total travel time per iteration to reach a particular destination
   const waitTimesForDest = new Uint8Array(nMinutes) // wait time per iteration for particular destination
@@ -67,7 +72,7 @@ export default function propagate ({ query, stopTreeCache, origin, callback }) {
         }
       }
 
-      callback({
+      next({
         travelTimesForDest,
         walkTimesForDest,
         inVehicleTravelTimesForDest,
