@@ -13,11 +13,11 @@ import {create as createStopTreeCache} from './stop-tree-cache'
 module.exports = createHandler({
   putGrid (ctx, message) {
     if (!ctx.grids) ctx.grids = new Map()
-    ctx.grids.set(message.key, createGrid(message.grid))
+    ctx.grids.set(message.id, createGrid(message.grid))
   },
   accessibilityForGrid (ctx, message) {
     return accessibilityForGrid({
-      grid: ctx.grids.get(message.grid),
+      grid: ctx.grids.get(message.gridId),
       cutoff: message.cutoff,
       surface: ctx.surface
     })
@@ -46,7 +46,7 @@ module.exports = createHandler({
   },
   generateSurface (ctx, message) {
     const { spectrogramData, ...surface } = getSurface({
-      grid: ctx.grids.get(message.grid),
+      grid: ctx.grids.get(message.gridId),
       origin: ctx.origin,
       query: ctx.query,
       stopTreeCache: ctx.stopTreeCache,
@@ -57,8 +57,8 @@ module.exports = createHandler({
 
     return { spectrogramData }
   },
-  generateDestinationData (ctx, message) {
-    const { to, from } = message
+  generateDestinationData (ctx, message, log) {
+    const {to, from} = message
 
     const travelTime = ctx.surface.surface[to.y * ctx.query.width + to.x]
     const waitTime = ctx.surface.waitTimes[to.y * ctx.query.width + to.x]
@@ -75,6 +75,7 @@ module.exports = createHandler({
         to
       }),
       transitive: getTransitiveData({
+        log,
         network: ctx.transitiveNetwork,
         origin: ctx.origin,
         from,
